requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.12.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.2.4
psycopg2-binary==2.9.3
python-decouple==3.6
flake8==4.0.1
django-redis==5.2.0
model-bakery==1.5.0
pytest-django==4.5.2